+++ maple2latex/src/maple_tokenize.py
@@ -21,6 +21,8 @@ symbols = [
     ")",
     "{",
     "}",
+    "[",
+    "]",
     "\'",
     "\"",
     "|",
